Begin getting rid of callbacks returning strings
Strings were never a good choice to begin with.  However, with the efforts
to improve translatability, they became even worse.  This is an attempt to
clean up the current mess and provide a more structured way for modules to
commmunicate with the shell.

The Computer module has been partially converted and changes will be
provided next.  It's partially converted since some of its information is
still stored as strings.

The Shell still only understands strings; that's why the structured data is
converted to a string the way it is.  Once all modules are converted, the
Shell can be modified to handle struct Info directly without having to parse
the .ini-like strings.