Remove the appended "Module" from about module dialog title
The string is constructed one word at a time, "About", X, "Module"
and it ends up being very strange in languages other than English.
Cutting out the module makes it a little better, and not much is
lost. A module could add "Module" to it's title, and then at
least that string can be translated as a complete unit.

Signed-off-by: Burt P <pburt0@gmail.com>