devices: Remove product-version from string returned by getMotherboard()
This reverts a change 2db563687071c099851c59396bdde29a00dba156.

It seems to me that using product version in this way gives
inconsistent values.
The recent benchmark result submitted demonstrates the problem:
`2.0 / X370 SLI PLUS (MS-7A33) (Micro-Star International Co., Ltd.)`

The product version `2.0` is put in front, but without the product
name or product vendor for context, it doesn't really add anything.
The board vendor, which otherwise would be in front, is now in () at
the end.

When the product version is not defined, then the motherboard is
reported as `<board vendor> <board name>`. The only time it is
`<board name> (<board vendor>)` is when product version is defined.

If this change is applied, that same motherboard would be reported
`Micro-Star International Co., Ltd. X370 SLI PLUS (MS-7A33)`
which, to me, makes more sense for getMotherboard(). I have no idea
what "product" this is version 2.0 of.

Signed-off-by: Burt P <pburt0@gmail.com>