params: html_ok -> markup_ok
The html_ok param added in
970174b0897d40b804808632784ffa1544d9da93
is misleading. It turns out that it is not HTML,
but Pango Markup that is used by GTK, and a subset of that
happens to also work as HTML4 for the HTML report generation.

This change renames html_ok to markup_ok in ProgramParameters
and adds a comment explaining what common set of tags may be
used.

Also, if report generation happens as a fallback after GUI
initialization failed, then also disable markup, unless report
will be generated as HTML.

Pango markup info at:
https://developer.gnome.org/pango/stable/PangoMarkupFormat.html

Signed-off-by: Burt P <pburt0@gmail.com>