GTK3/Benchmarks disable guibench only
GTK3 build inadvertably disabled all benchmarking.
At present, the GTK3 build intentionally avoids building of "guibench"
due to non-trivial changes from GTK2 to GTK3.
However, all other benchmarks were disabled as well.
This was not by design, as non-gui benchmarks have no GTK3
dependencies.
The reason benchmarks were diasbled (did not show up in the menus)
was that the code for "benchark_gui" still referenced the missing "guibench"
routine. This routine was not present in the "benchmark.so", so the library
would not load due to the missing entry point.

This patch addresses the issue. All benchmarks (except guibench)
show/work fine with GTK3.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>