Use DetailView for SHELL_VIEW_NORMAL and the more info display
This change has been made mostly because we need more control in how
things are displayed, and fighting GTK+ so that this is possible isn't
going to cut it.

For instance, it is now possible to display links -- albeit only
external links at the moment, but internal links will be possible,
allowing for instance for one to click in a kernel module dependency to
know more about that particular module.  Links will also be useful to
go to the vendor website, and maybe other things (like web search for
Vendor + Product ID) and whatnot.

Eventually this will replace the summary view altogether.  That code
has a lot of drawbacks, including not being able to have its contents
copied to the clipboard, appear in reports, or have auto-updating
fields.  It stays for now, but it will be gone before 0.6 is released.

I had forgotten how messy this code is.  One of the reasons I lost the
interest in this project...  I want to hack on a nicer code base but I
also lack the strong motivation to write the shell from the scratch.