Add script to convert benchmark.conf to benchmark.json
The new server will send and retrieve information in JSON, as that was
what was easier to do.  Having a more verbose format makes things a
whole lot easier to work with, too, so there's this bonus.

This script replicates some of the work that HardInfo does to adapt the
various versions of benchmark.conf, including cleaning up the CPU names
and whatnot.  (Might be a good idea to store the raw values in the
database too just in case.)

Results seem sane from a first glance, but this will need some testing.

The idea is to use this script once, use it to rewrite the benchmark
result loader, and then populate the server with this information.  After
that, this script can be thrown away.