Revert "Applies SVG graphics to the tooltip (note)"
This reverts commit 92a53ca7837ea9e5b08555756c7441f179eaa166.

@tutralex
This does not work!! nor look good!! - the note is no longer a note
(does not standout as highlighted informational area that can be closed)
as you have removed the box making it a highlighted note and made close
just an icon to the lower right.
This is not solving any of the mentioned problems.